// ui/src/lib/api/model/sessionEvent.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform that provides scalable browser sessions and worker pool management.
 * OpenAPI spec version: 1.0
 */
import type { SessionEventData } from './sessionEventData';
import type { SessionEventType } from './sessionEventType';

/**
 * Session event with type, data and timestamp
 */
export interface SessionEvent {
  data?: SessionEventData;
  event?: SessionEventType;
  id?: string;
  session_id?: string;
  timestamp?: string;
}